refactor(llvm): split compiled polling tests
